Put all meter provider tests in one file
